add basic templating (css/js broken)
