--- portfolio/src/main/resources/templates/index.html
@@ -1,150 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-	<meta charset="utf-8">
-	<title>Mudit Gupta</title>
-	<link rel="icon" type="image/png" sizes="32x32" href="static/images/logo.png">
-	<link rel="stylesheet" type="text/css" href="static/styles/base.css">
-	<link rel="stylesheet" type="text/css" href="static/styles/header.css">
-	<link rel="stylesheet" type="text/css" href="static/styles/navbar.css">
-	<link rel="stylesheet" type="text/css" href="static/styles/konami.css">
-	<link rel="stylesheet" type="text/css" href="static/styles/home.css">
-	<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-</head>
-
-<body>
-	<header>
-		<div id="top_nav" class="nav_bar">
-			<img class="logo" src="static/images/logo.png" alt="" onclick="navToSection('#top')">
-			<a class="nav_bar_item" id="nav_who_am_i"
-				onclick="navToSection('#who_am_i')"><i class="fa fa-fw fa-user"></i> Who Am I?</a>
-			<a class="nav_bar_item" id="nav_projects"
-				onclick="navToSection('#projects')"><i class="fa fa-fw fa-book"></i> Projects</a>
-			<a class="nav_bar_item" id="nav_gallery" onclick="navToSection('#gallery')"><i class="fa fa-fw fa-book"></i>
-				Gallery</a>
-			<a class="nav_bar_item" id="nav_find_me_online"
-				onclick="navToSection('#find_me_online')"><i class="fa fa-fw fa-globe"></i> Find Me Online</a>
-			<a class="nav_bar_item" id="nav_comments" onclick="navToSection('#comments')"><i class="fa fa-fw fa-file"></i>
-				Comment</a>
-			<a class="menu" href="javascript:void(0);" onclick="responsiveNav()"><i class="fa fa-bars"></i></a>
-		</div>
-		<div class="title_block" id="top">
-			<img id="title_bkgd_img" src="static/images/homies low res.png" alt="blur friends"/>
-			<div class="title_block_inner">
-				<img id="title_image" src="static/images/homies low res.png" alt="friends"/>
-				<div class="title_content">
-					<h1 class="title_content_item">
-						<center id="title">Mudit Gupta</center>
-					</h1>
-					<div class="title_content_item" id="quick_links">
-					</div>
-				</div>
-			</div>
-		</div>
-	</header>
-	<div class="mainBody">
-		<div class="mainContent">
-			<div class="content_block" id="who_am_i">
-				<h1>Who Am I?</h1>
-				<div class="content">
-					<p class="about-me-text">
-						<img class="inline-pic-left about-me-pic" src="static/images/im cool.JPG" />
-            Hey! I'm Mudit Gupta. I was born in Georgia on March 17, 2001, and I've been living in the suburbs my whole life.
-            I grew up with the gift of my parents having hundreds of friends. All the family around me molded me into the
-            fun-loving person I am today. I bring my goofy (slightly immature) antics into every room I enter, and as I got older,
-            I realized that the smiles I could put on other people's faces put a smile twice as big on mine. My passion for making
-            others happy has driven me throughout my life. When I'm not working on something of my own, I often find myself helping
-            someone else work on something of their own.
-          </p>
-          <p class="about-me-text">
-            <img class="inline-pic-right about-me-pic" src="static/images/senior pic.JPG" />
-            In highschool, my enthusiasm for robotics became my personality. I spent all four years invested in my robotics team,
-            and we reached the international competition twice! After graduating from GSMST, I became a student at Georgia Tech.
-            I'm working towards a Computer Science degree, focusing on artificial intelligence and modeling/simulation.
-          </p>
-				</div>
-			</div>
-			<div class="content_block" id="projects">
-				<h1>Projects</h1>
-				<div class="content">
-					<ul>
-						<li>VEX Robotics 2016-2019</li>
-						<li>Woke! app</li>
-						<li>Pineapple game</li>
-						<li>CS-Vis</li>
-					</ul>
-				</div>
-			</div>
-			<div class="content_block" id="gallery">
-				<h1>Gallery</h1>
-				<div class="content" id="gallery-container">
-					<img src="static/images/IMG_5778.JPG" />
-					<img src="static/images/stuco retreat.jpeg" />
-        </div>
-      </div>
-      <div class="content_block" id="find_me_online">
-        <h1>Find Me Online</h1>
-        <div class="content" id="badge-container">
-          <div class="LI-profile-badge" data-version="v1" data-size="medium" data-locale="en_US"
-              data-type="horizontal" data-theme="light" data-vanity="muditg317" data-uid="173090">
-            <a class="LI-simple-link" href="https://www.linkedin.com/in/muditg317?trk=profile-badge">
-              Mudit Gupta
-            </a>
-          </div>
-          <div class="github-card" data-github="muditg317" data-width="400" data-height=""
-              data-theme="default"></div>
-        </div>
-      </div>
-      <div class="content_block" id="comments">
-        <h1>Comments</h1>
-        <div class="content" id="comment-container">
-          <form onsubmit="postComment(); return false">
-            <input type="text" id="comment-text" />
-            <input type="submit" value="Post comment!" />
-            <input onchange="loadComments()" type="number" id="num-comments" name="numComments" min="1" max="10000" required="" value="100">
-          </form>
-          <div id="comment-list"></div>
-          <button onclick="deleteComments()">Delete Comments</button>
-        </div>
-      </div>
-    </div>
-    <div class="rising_button" id="back_to_top_block" onclick="goToTop()">
-      <p>Back to top:</p>
-      <img src="static/images/up_arrow.png" alt="">
-    </div>
-  </div>
-  <div class="konamiGame">
-    <div class="konami" id="bird">
-    </div>
-    <div class="konami pipe" id="pipeBottom">
-    </div>
-    <div class="konami pipe" id="pipeTop">
-    </div>
-    <div class="konami score">
-      <p id="score">0</p>
-    </div>
-  </div>
-  <div class="pongGame">
-    <div class="pong" id="ball">
-    </div>
-    <div class="pong pipe" id="user">
-    </div>
-    <div class="pong pipe" id="cpu">
-    </div>
-    <div class="pong score">
-      <p id="score">0</p>
-    </div>
-  </div>
-  <script src="static/scripts/base.js" charset="utf-8"></script>
-  <script src="static/scripts/navbar.js" charset="utf-8"></script>
-  <script src="static/scripts/toTop_button.js" charset="utf-8"></script>
-  <script src="static/scripts/main.js" charset="utf-8"></script>
-  <script src="static/scripts/comments.js" charset="utf-8"></script>
-  <script src="static/scripts/konami.js" charset="utf-8"></script>
-  <script type="text/javascript" src="https://platform.linkedin.com/badges/js/profile.js" async defer>
-  </script>
-  <script src="//cdn.jsdelivr.net/github-cards/latest/widget.js"></script>
-</body>
-
-</html>
