+++ portfolio/src/main/resources/templates/title_block.html
@@ -1,0 +1,19 @@
+{% set styles = styles.append(
+    "static/styles/title_block.css",
+ ) %}
+<div class="title_block" id="top">
+  <img id="title_bkgd_img" src="static/images/homies low res.png" alt="blur friends"/>
+  <div class="title_block_inner">
+    <img id="title_image" src="static/images/homies low res.png" alt="friends"/>
+    <div class="title_content">
+      <h1 class="title_content_item">
+        <center id="title">Mudit Gupta</center>
+      </h1>
+      <div class="title_content_item" id="quick_links">
+      </div>
+    </div>
+  </div>
+</div>
+{% set scripts = scripts.append(
+    "static/scripts/title_block.js"
+  ) %}
